Colocate more Room types
